+++ webapp/src/network/mutations.js
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag'
 
 export const CREATE_TRANSACTION = gql`
-  mutation CreateTransaction ($amount: Float!, $credit: Boolean!, $debit: Boolean!, $description: String!, $userId: ID!, $merchantId: ID!) {
+  mutation CreateTransaction ($amount: Integer!, $credit: Boolean!, $debit: Boolean!, $description: String!, $userId: ID!, $merchantId: ID!) {
     createTransaction(amount: $amount, credit: $credit, debit: $debit, description: $description, userId: $userId, merchantId: $merchantId) {
       id
       amount
@@ -13,7 +13,7 @@ export const CREATE_TRANSACTION = gql`
 `
 
 export const EDIT_TRANSACTION = gql`
-  mutation UpdateTransaction ($id: ID!, $amount: Float!, $credit: Boolean!, $debit: Boolean!, $description: String!, $userId: ID!, $merchantId: ID!) {
+  mutation UpdateTransaction ($id: ID!, $amount: Integer!, $credit: Boolean!, $debit: Boolean!, $description: String!, $userId: ID!, $merchantId: ID!) {
     updateTransaction(id: $id, amount: $amount, credit: $credit, debit: $debit, description: $description, userId: $userId, merchantId: $merchantId) {
       id
       amount
